requirements.txt
pika==1.3.1
kombu==5.2.4
python-socketio==5.8.0
python-dotenv==1.0.0